Fix char_unescape_table lookup cast

+++ include/glaze/json/read.hpp
@@ -758,7 +758,7 @@ namespace glz
                      }
                      else {
                         p += next;
-                        *p = char_unescape_table[*start];
+                        *p = char_unescape_table[uint8_t(*start)];
                         if (*p == 0) [[unlikely]] {
                            ctx.error = error_code::invalid_escape;
                            return;
@@ -915,7 +915,7 @@ namespace glz
                            }
                            else {
                               p += next;
-                              *p = char_unescape_table[*start];
+                              *p = char_unescape_table[uint8_t(*start)];
                               if (*p == 0) [[unlikely]] {
                                  ctx.error = error_code::invalid_escape;
                                  return;
